Moved all the models and restapi classes to the data package
