+++ app/src/main/java/com/thesunnahrevival/sunnahassistant/data/model/Reminder.java
@@ -1,4 +1,4 @@
-package com.thesunnahrevival.sunnahassistant.data;
+package com.thesunnahrevival.sunnahassistant.data.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
