+++ app/src/main/java/com/thesunnahrevival/sunnahassistant/data/remote/GeocodingInterface.java
@@ -1,4 +1,6 @@
-package com.thesunnahrevival.sunnahassistant.restapi;
+package com.thesunnahrevival.sunnahassistant.data.remote;
+
+import com.thesunnahrevival.sunnahassistant.data.model.GeocodingData;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
