+++ app/src/main/java/com/thesunnahrevival/sunnahassistant/views/interfaces/ReminderItemInteractionListener.java
@@ -3,7 +3,7 @@ package com.thesunnahrevival.sunnahassistant.views.interfaces;
 import android.view.View;
 import android.widget.CompoundButton;
 
-import com.thesunnahrevival.sunnahassistant.data.Reminder;
+import com.thesunnahrevival.sunnahassistant.data.model.Reminder;
 
 public interface ReminderItemInteractionListener {
     void onToggleButtonClick(CompoundButton buttonView, boolean isChecked, Reminder reminder);
