+++ app/src/main/res/layout/content_main.xml
@@ -8,7 +8,7 @@
 
         <import type="android.view.View" />
 
-        <import type="com.thesunnahrevival.sunnahassistant.data.Reminder" />
+        <import type="com.thesunnahrevival.sunnahassistant.data.model.Reminder" />
 
         <variable
             name="viewmodel"
@@ -16,7 +16,7 @@
 
         <variable
             name="nextReminder"
-            type="com.thesunnahrevival.sunnahassistant.data.Reminder" />
+            type="com.thesunnahrevival.sunnahassistant.data.model.Reminder" />
 
     </data>
