+++ app/src/main/res/layout/reminder_card_view.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="reminder"
-            type="com.thesunnahrevival.sunnahassistant.data.Reminder" />
+            type="com.thesunnahrevival.sunnahassistant.data.model.Reminder" />
     </data>
 
     <androidx.cardview.widget.CardView
