+++ app/src/main/res/layout/reminder_details_bottom_sheet.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="reminder"
-            type="com.thesunnahrevival.sunnahassistant.data.Reminder" />
+            type="com.thesunnahrevival.sunnahassistant.data.model.Reminder" />
 
         <variable
             name="isNew"
